# feat: significantly optimize the time consumption of clip vision