@huchenlei 2024-01-15
I think this closes https://github.com/Mikubill/sd-webui-controlnet/issues/2436. Thanks for the fix!